+++ baseapp/state.go
@@ -2,11 +2,13 @@ package baseapp
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	abci "github.com/tendermint/tendermint/abci/types"
 )
 
 type state struct {
-	ms  sdk.CacheMultiStore
-	ctx sdk.Context
+	ms           sdk.CacheMultiStore
+	ctx          sdk.Context
+	eventHistory []abci.Event
 }
 
 // CacheMultiStore calls and returns a CacheMultiStore on the state's underling